Add solution for problem `A - Game 23` Link: https://codeforces.com/contest/1141/problem/A Tags: implementation, math, *1000 Ref: https://codeforces.com/contest/1141/submission/51499853

+++ codeforces/1141/A.cpp
@@ -26,12 +26,12 @@ int main()
 
     while(q.empty()!=1)
     {
-        ll x=q.front();
+        int x=q.front();
         q.pop();
 
-        for(ll i=2; i<=3; i++)
+        for(int i=2; i<=3; i++)
         {
-            ll y=x*i;
+            int y=x*i;
             if(visited[y]==0 && y<=des)
             {
                 visited[y]=1;
